+++ src/core/store.js
@@ -1,10 +1,10 @@
 import { configureStore } from "@reduxjs/toolkit";
 import createSagaMiddleware from "redux-saga";
 import rootSaga from "./rootSaga";
-import movieListReducer from './features/movieListSlice';
-import peopleListReducer from './features/peopleListSlice';
-import movieDeailsReducer from './features/movieSlice';
-import personDeailsReducer from './features/personSlice';
+import movieListReducer from '../features/movieListSlice';
+import peopleListReducer from '../features/peopleListSlice';
+import movieDeailsReducer from '../features/movieSlice';
+import personDeailsReducer from '../features/personSlice';
 
 const sagaMiddleware= createSagaMiddleware();
